reformat repo
store rounder code in a folder called rounder, which can just be imported. benefits of this is more separated code and the ability to have rounder not just be a single python file. the option to have it be a single file is still obtainable of course of the user wishes so